--- src/hooks/registry.ts
@@ -1,57 +1,0 @@
-/*
- * Copyright (C) 2022 PixieBrix, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { RegistryId } from "@/core";
-import { RecipeDefinition } from "@/types/definitions";
-import { useSelector } from "react-redux";
-import { selectAllRecipes } from "@/recipes/recipesSelectors";
-import { useMemo } from "react";
-
-export type StateSelector<T> = {
-  /**
-   * The value, or `undefined` if the state is loading or there was an error computing the state
-   */
-  data: T | undefined;
-
-  /**
-   * True if the async state is loading
-   */
-  isLoading: boolean;
-};
-
-/**
- * Lookup a recipe from the registry by ID
- */
-export function useRecipe(id: RegistryId): StateSelector<RecipeDefinition> {
-  const { data: allRecipes, isLoading } = useAllRecipes();
-  const recipe = useMemo(
-    () =>
-      allRecipes?.length
-        ? allRecipes.find((x) => x.metadata.id === id)
-        : undefined,
-    [id, allRecipes]
-  );
-
-  return { data: recipe, isLoading };
-}
-
-/**
- * Pulls all recipes from the registry
- */
-export function useAllRecipes(): StateSelector<RecipeDefinition[]> {
-  return useSelector(selectAllRecipes);
-}
